+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Enable logarithmic scale for the residual calculation (non filtered) from the paper: BRDF Models for Accurate and
   Efficient Rendering of Glossy Surfaces
 - Support for saving height field files in EXR image format: single channel 32-bit float with extra surface information
+- Implement cargo like external subcommands
 
 ## 0.3.2 - 2024-08-16